requirements-dev.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
python-dateutil>=2.8.0
chardet>=5.0.0
fastjsonschema>=2.18.0

